refactor(sales-app): use BigDecimal whenever the product fee or discount is calculated

+++ sales-app/app/src/main/java/com/peyess/salesapp/data/adapter/edit_service_order/payment_fee/PaymentFeeDocumentAdapter.kt
@@ -7,6 +7,6 @@ fun PaymentFeeDocument.toEditPaymentFeeEntity(): EditPaymentFeeEntity {
     return EditPaymentFeeEntity(
         saleId = saleId,
         method = method,
-        value = value,
+        value = value.toDouble(),
     )
 }
